changes psp / istio w/o cni predicate conditions
even when the PSP API is removed from k8s, this rule addition will not
cause an error. if istio is running w/o the CNI plugin installed, then
we need to bind to a PSP. if psps are enabled, then we'll need to bind
to a "restricted" policy. so changing this to an OR condition for now.