Polish Helm chart (#8)
* condenses helm chart

removes buildkit subchart. given that this chart and the parent need to
share templates, flag variables, etc, this original structure results in
a coupled and ugly form.

* adds network policies for controller and buildkit

enabled by default

* uses nonroot distroless image for controller

* tightens pod execution security

- adds psp and rbac generator for buildkit
- condenses config values for netpols and psps
- removes token from buildkit sa
- runs vector log processor as nonroot user

* makes changes to support istio with and without cni

* optional rendering of buildkit replicas field

this field is not rendered when the value is 0. this prevents upgrades
from overwriting the current replicat count set by the controller and
unintentionally scaling the cluster to 0.

* unifies naming scheme for global rbac resources

* fixes typo in crd-hooks.yaml file

* sets missing storageClassName in buildkit sts

* changes istio.cniPluginInstalled to istio.cni

* removes supportsSeccompGA check

- minimum k8s version supported is v1.19
- enforced via Chart.yaml

* removes podSecurityContext field from buildkit

we really don't want folks messing with this and we can always add it
back later "if" necessary.

also sets "fsGroup" to same UID/GID as the user when running rootless.
this gives the user the permission to write to pvc provisioned by block
storage SCs.

* quotes amqp url value in controller secret config

missed this earlier. special chars in the password can cause the YAML
parser to barf

* adds support for custom CA configmap vol

* changes psp / istio w/o cni predicate conditions

even when the PSP API is removed from k8s, this rule addition will not
cause an error. if istio is running w/o the CNI plugin installed, then
we need to bind to a PSP. if psps are enabled, then we'll need to bind
to a "restricted" policy. so changing this to an OR condition for now.