changes default status transitions
setting "Created" transition to processed so that a status update is not
sent for this initial phase. this setup is really a workaround so that
we can patch the status when messages are sent. if the json path does
not already exist, then the patch will silently fail

hopefully we can figure out a better way to handle this in the future