adds ListWorkers check when building buildkit client
a new buildkit client does not verify its connection with a buildkitd
worker when first created so a garbage addr like "foo.bar:1234" will
not return an error. running a "list workers" operation like we do in
the statefulset readiness check for buildkit should solve this problem
and also guarantee that the TCP server has had time to set up its TLS
certificates.

retrying the operation with a backoff makes several attempts to
establish the connection before the client is returned to the caller