adds .Values.uninstallCRDs to helm chart (#73)
by default, this chart will no longer uninstall the CRDs when the helm
chart release is deleted. this will prevent the accidental deletion of
CRs that might be needed and also, in the case of uninstall/re-install,
keep k8s from losing its mind whenever the CRDs are removed and CRs are
present. this latter case usually results in the k8s CP taking a few
mins to figure out what's happening and stabilize.