sets buildkit workerpool to 0 replicas by default
if/when we perform a helm upgrade that affects either buildkit or the
controller, then either the buildkit sts will scale down or the manager
will terminate and all of the builds will be terminated. hence, there is
no risk with setting the replicas to 0 here.

also added a TODO to improve the buildkit worker pool