updates buildkit workerpool
the previous "abnormal" descriptor didn't really encompass the details
of what's happening in k8s. it's possible that pod can be Running but
not ready, in which it is "starting" and should be afforded an
opportunity to finish the process.

furthermore, when performing math, it makes more sense to have the
subtractions variable declared before the loop. this allows us to gather
pods into groups like "starting" or "pending" for display purposes but
still be able to apply offsets for certain conditions using the
subtractions var