Complete GKE testing (#75)
* adds istio to function tests

* adds more functional test data

and uses prescribed "testdata" directory instead of "fixtures"

* fixes build-arg validation

we only need to ensure that a single "=" char exists between the key and
value, not if the value has equal signs in it

* build-arg validation accounts for blank "key"

* adds many many many gke tests

* adds more gke tests

- fixes teardown
- extracts tasks into reusable helper funcs
- adds test funcs for checking log and message delivery

* adds mutli-stage build context

* adds multi-stage and concurrent tests

* updates tooling

* updates buildkit and cobra deps

* updates buildkit workerpool

the previous "abnormal" descriptor didn't really encompass the details
of what's happening in k8s. it's possible that pod can be Running but
not ready, in which it is "starting" and should be afforded an
opportunity to finish the process.

furthermore, when performing math, it makes more sense to have the
subtractions variable declared before the loop. this allows us to gather
pods into groups like "starting" or "pending" for display purposes but
still be able to apply offsets for certain conditions using the
subtractions var

* refactors buildkit worker pool scale determination

YAWPR (yet another worker pool refactor)

trying to collect builder states and determine the proper statefulset
scale along with servicing build requests resulted in a BRUTAL for loop
inside the worker pool. furthermore, there are so many edge cases that
this required special "offsets" to get all test cases to pass. this
change extracts that logic into a dedicated type that focuses singularly
on evaluating pod states and, as a later step, calculates the proper
replica count

additionally, this improves the worker pool tests so that we test
behaviors more thoroughly by adding more tests with requests > 0

* adds DSE docker-context

* updates buildkit storage size for DSE

* adds dse test

and updates createBuild timeout to account for this gargantuan image
build time

* updates testenv dep

adds bigger gke nodes so we don't have to scale up the cluster during
testing

* vector ignores empty messages during testing

so we can validate the payload sent to redis w/o erroring out since some
JSON payloads were sent with an empty "msg" field. this is not an error,
the vector config just needed tweaking

* adds log field to ScaleArbiter

and logs out valuable state information when invoking DetermineReplicas

* updates buildkit image tag

to align with library upgrade

* adds note to garbageClientHack

so the peoples understand why