Fix worker pool (#126)
* DOM-52319: Correct AutoscalingPool

Start only returns when the pool is "done".
Close is unnecessary since it is never used.

When the result channel is closed, the pool is closed.

* PLAT-7635: Mark active ImageBuilds as failed on restart

When starting up, any "active" (initializing, running) ImageBuild is now
failed. The buildkit daemon behind it is being terminated.

* Update dependencies

* Re-enable leader election

There must be only one active instance.