v0.11.11 cedits (#316)
* Update to use DockerAuthProviderConfig struct

The function parameters have changed to use a single struct instead
of a configFile and the tlsparameters like before. Updated this to
follow the new format

* Add debug endpoint

For debugging purposes adding the debug endpoint so we can see
if endpoints are registered correctly and are served

* Upgrade to golang-ci-lint v2

* Fix lint errors

* Update pre-commit config

* Update Golang version

* Fix Webhook

The new controller-runtime uses new defaulters and validators which
breaks the old format for registering the Webhook. This PR resolves
that.

* Allow debug builds of the Dockerfile

The current distroless image makes it difficult to debug the image
if there is an issue. Adding a debug dockerfile, which installs
all the needed tools for debugging locally as well an in a cluster
Debug builds should not be deployed on a prod cluster.

* review comments

* Improve Docker build caching and speed

Run docker steps in parallel to improve build times