separate caches for each build
According to the docs [1]

If you build multiple images, each build will overwrite the cache of the previous
leaving only the final cache. To preserve the cache for multiple builds, you
can specify this scope attribute with a specific name.

Since the cache names are the same the debug cache is being used for the prod one
leading to weirdness

[1]:https://docs.docker.com/build/cache/backends/gha/#scope